python-server/requirements.txt
flask
google-cloud-vision